youtube/pyproject.toml
[project]
name = "youtube"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.38.0",
    "yt-dlp>=2024.8.6",
    "openai>=1.45.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
